no errors, hover effects only on link, reverted to commit with working redirects
