+++ src/components/Classroom.js
@@ -9,7 +9,7 @@ class Classroom extends Component {
     render() {
         return (
             <div className="Classroom">
-                here is stuff
+                here is classroom stuff
             </div>
         )
     }
