+++ src/styles/_variables.scss
@@ -6,3 +6,4 @@ $mint: RGBA(78, 201, 176, 1.00);
 $shadow: RGBA(20, 20, 20, .4);
 $gray: #1c1c1c;
 $white: whitesmoke;
+$gridColumnWidth: 123px;
