+++ src/styles/componentStyles/_Classroom.scss
