+++ src/styles/componentStyles/_Home.scss
@@ -1,6 +1,6 @@
 .Home {
     display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr;
+    // grid-template-columns: 1fr 1fr 1fr 1fr;
     grid-template-rows: 100%;
     height: 100vh;
     width: 100vw;
@@ -30,7 +30,7 @@
             size: 2.1rem;
             weight: 200;
         }
-        transition: all .7s;
+        transition: all .5s;
         box-shadow: 0 0 0 7px RGBA(245, 245, 245, 0);
         
         &.Home_classroom {
@@ -45,10 +45,24 @@
             grid-column: 4 / span 1;
             background-color: $DMblue;
         }
-        &.Home_classroom:hover, &.Home_extra_resources:hover, &.Home_provo:hover {
-            background-color: $white;
-            box-shadow: 0 10px 20px 7px RGBA(30, 30, 30, 1);
-            color: $gray;
-        }
+        
     }
-}
+    .Home_login .linkBar            { background-color: $gray; }
+    .Home_classroom .linkBar        { background-color: $darkbrick; }
+    .Home_extra_resources .linkBar  { background-color: $brick; }
+    .Home_provo .linkBar            { background-color: $DMblue; }
+
+    .linkBar {
+        width: $gridColumnWidth;
+        height: 100vh;
+        display: inline-block;
+        transition: .5s;
+    }
+    .linkBar:hover {
+        background-color: $white;
+        box-shadow: 0 10px 20px 7px RGBA(30, 30, 30, 1);
+        color: $gray;
+    }
+} // end of .Home
+
+
