--- src/styles/styleObjects/spaces.js
@@ -1,16 +1,0 @@
-export default {
-    spaceBar: {
-        collapsed: {
-            transform: 'rotate(90deg)'
-            ,transformOrigin: '30px 30px'
-            // ,color: 'whitesmoke'
-            // ,transformOrigin: 'left top'
-            // ,writingMode: 'vertical-lr'
-            // ,textOrientation: 'upright'
-        }
-        ,
-        expanded: {
-            color: 'blue'
-        }
-    }
-}
